+++ bin/faucet/frontend/app.js
@@ -1,5 +1,6 @@
 import { MidenWalletAdapter } from "@demox-labs/miden-wallet-adapter-miden";
-import { PrivateDataPermission, WalletAdapterNetwork } from "@demox-labs/miden-wallet-adapter-base";
+import { AllowedPrivateData, PrivateDataPermission, WalletAdapterNetwork } from "@demox-labs/miden-wallet-adapter-base";
+import { Endpoint, NoteId, RpcClient } from "@demox-labs/miden-sdk";
 import { Utils } from './utils.js';
 import { UIController } from './ui.js';
 import { getConfig, getMetadata, getPowChallenge, getTokens, get_note } from "./api.js";
@@ -10,6 +11,7 @@ export class MidenFaucetApp {
         this.tokenAmountOptions = [100, 500, 1000];
         this.metadataInitialized = false;
         this.apiUrl = null;
+        this.rpcClient = null;
         this.baseAmount = null;
         this.powLoadDifficulty = null;
 
@@ -20,7 +22,9 @@ export class MidenFaucetApp {
         }
 
         this.walletAdapter = new MidenWalletAdapter({ appName: 'Miden Faucet' });
-
+        this.walletAdapter.on('accountChange', (_) => {
+            this.walletAdapter.disconnect();
+        });
         this.init();
     }
 
@@ -28,6 +32,7 @@ export class MidenFaucetApp {
         try {
             let config = await getConfig();
             this.apiUrl = config.api_url;
+            this.rpcClient = new RpcClient(new Endpoint(config.node_url));
             this.setupEventListeners();
             this.startMetadataPolling();
         } catch (error) {
@@ -46,7 +51,7 @@ export class MidenFaucetApp {
 
     async handleWalletConnect() {
         try {
-            await this.walletAdapter.connect(PrivateDataPermission.UponRequest, WalletAdapterNetwork.Testnet);
+            await this.walletAdapter.connect(PrivateDataPermission.Auto, WalletAdapterNetwork.Testnet, AllowedPrivateData.None);
 
             if (this.walletAdapter.accountId) {
                 this.ui.setRecipientAddress(this.walletAdapter.accountId);
@@ -83,9 +88,14 @@ export class MidenFaucetApp {
             const nonce = await this.findValidNonce(powData.challenge, powData.target);
 
             this.ui.updateMintingTitle('MINTING TOKENS');
-            this.ui.updateProgressBar(50);
+            this.ui.updateProgressBar(33);
 
             const getTokensResponse = await getTokens(this.apiUrl, powData.challenge, nonce, recipient, amount, isPrivateNote);
+
+            this.ui.updateMintingTitle('CONFIRMING TRANSACTION');
+            this.ui.updateProgressBar(66);
+
+            await this.pollNote(getTokensResponse.note_id);
 
             this.ui.showCompletedModal(
                 recipient,
@@ -190,6 +200,54 @@ export class MidenFaucetApp {
             console.error('Error downloading note:', error);
             this.ui.showPrivateModalError('Failed to download note: ' + error.message);
         }
+    }
+
+    pollNote(noteId) {
+        return new Promise((resolve, reject) => {
+            // Poll every 500ms for the first 10 seconds, then every 1s for the next 30 seconds, then every 5s.
+            let currentInterval = 500;
+            let pollInterval;
+            let elapsedTime = 0;
+            // Timeout after 5 minutes
+            const timeout = 300000;
+            let timeoutId;
+
+            const poll = async () => {
+                try {
+                    const note = await this.rpcClient.getNotesById([NoteId.fromHex(noteId)]);
+                    if (note && note.length > 0) {
+                        clearInterval(pollInterval);
+                        clearTimeout(timeoutId);
+                        resolve();
+                        return;
+                    }
+
+                    elapsedTime += currentInterval;
+
+                    if (elapsedTime <= 10000) {
+                        currentInterval = 500;
+                    } else if (elapsedTime <= 40000) {
+                        currentInterval = 1000;
+                    } else {
+                        currentInterval = 5000;
+                    }
+
+                    // Update the interval
+                    clearInterval(pollInterval);
+                    pollInterval = setInterval(poll, currentInterval);
+                } catch (error) {
+                    console.error('Error polling for note:', error);
+                    clearInterval(pollInterval);
+                    clearTimeout(timeoutId);
+                    reject('Error fetching note confirmation.');
+                }
+            };
+            pollInterval = setInterval(poll, currentInterval);
+            timeoutId = setTimeout(() => {
+                clearInterval(pollInterval);
+                reject(new Error('Timeout while waiting for tx to be committed. Please try again later.'));
+            }, timeout);
+        });
     }
 
     async findValidNonce(challenge, target) {